fix: array null key access is deprecated
When an object was hydrated using a typehint docblock it can happen that it tries to check if the sub-array key exists in order to know if it should traverse further down the object tree.

When an object is hydrated using a type-hinted docblock, it may attempt to check whether a sub-array key exists in order to determine if it should traverse further down the object tree.

Previously, this check relied on the null-coalesce operator to fall back to false. However, when the array is empty, array_key_first() returns null, and accessing a null offset is deprecated as of PHP 8.5. To address this, we added a count() check to detect empty arrays.